Add files via upload
This is a code for a basic pairs trading strategy. It uses a Indian Futures Stock prices as input and a text file having bifurcation of bifurcation of Future Stocks as per Industry they belong. It uses MySql to store the trades done where PnL is also calculated as the position exits.